Revert "trying to fix state issue"
This reverts commit 3406745a76119881b272b24916f062ee6bdbe48b.
